Moved availability_query_refactor.py to overwrite original availability_query.py file
