+++ src/hutch_bunny/core/solvers/availability_solver.py
@@ -1,11 +1,13 @@
-import pandas as pd
 from datetime import datetime
+
+import pandas as pd
 from dateutil.relativedelta import relativedelta
 from typing import Any, Callable, TypedDict, Union
 from sqlalchemy.sql.expression import ClauseElement
 from sqlalchemy import (
     CompoundSelect,
     or_,
+    and_,
     func,
     BinaryExpression,
     ColumnElement,
@@ -14,7 +16,7 @@ from sqlalchemy import (
     text,
     intersect,
     union,
-    literal,
+    literal
 )
 from hutch_bunny.core.db_manager import SyncDBManager
 from hutch_bunny.core.entities import (
@@ -23,8 +25,7 @@ from hutch_bunny.core.entities import (
     Measurement,
     Observation,
     Person,
-    DrugExposure,
-    ProcedureOccurrence,
+    DrugExposure
 )
 from tenacity import (
     retry,
@@ -34,7 +35,8 @@ from tenacity import (
     after_log,
 )
 
-from typing import Tuple
+from typing import Tuple 
+from sqlalchemy import exists
 
 from hutch_bunny.core.obfuscation import apply_filters
 from hutch_bunny.core.rquest_models.group import Group
@@ -42,7 +44,6 @@ from hutch_bunny.core.rquest_models.availability import AvailabilityQuery
 from sqlalchemy.engine import Engine
 from hutch_bunny.core.logger import logger, INFO
 
-from hutch_bunny.core.settings import Settings
 from hutch_bunny.core.rquest_models.rule import Rule
 import operator as op
 
@@ -58,25 +59,395 @@ class RuleTableQuery(TypedDict):
     inclusion: bool
 
 
-settings = Settings()
+class SQLDialectHandler:
+    @staticmethod
+    def get_year_difference(
+        engine: Engine,
+        start_date: ClauseElement,
+        birth_date: ClauseElement
+    ) -> ColumnElement[int]:
+        if engine.dialect.name == "postgresql":
+            return func.date_part("year", start_date) - func.date_part(
+                "year", birth_date
+            )
+        elif engine.dialect.name == "mssql":
+            return func.DATEPART(text("year"), start_date) - func.DATEPART(
+                text("year"), birth_date
+            )
+        else:
+            raise NotImplementedError("Unsupported database dialect")
 
 
-# Class for availability queries
-class AvailabilitySolver:
-    omop_domain_to_omop_table_map = {
-        "Condition": ConditionOccurrence,
-        "Ethnicity": Person,
-        "Drug": DrugExposure,
-        "Gender": Person,
-        "Race": Person,
-        "Measurement": Measurement,
-        "Observation": Observation,
-        "Procedure": ProcedureOccurrence,
-    }
+class OMOPRuleQueryBuilder:
+    """Builder for constructing OMOP queries from availability rules."""
+
+    def __init__(self, db_manager: SyncDBManager):
+        self.db_manager = db_manager
+        self.condition_query: Select[Tuple[int]] = select(ConditionOccurrence.person_id)
+        self.drug_query: Select[Tuple[int]] = select(DrugExposure.person_id)
+        self.measurement_query: Select[Tuple[int]] = select(Measurement.person_id)
+        self.observation_query: Select[Tuple[int]] = select(Observation.person_id)
+
+    def add_concept_constraint(self, concept_id: int) -> 'OMOPRuleQueryBuilder':
+        """Add standard concept ID constraints to all relevant tables."""
+        self.condition_query = self.condition_query.where(
+            ConditionOccurrence.condition_concept_id == concept_id
+        )
+        self.drug_query = self.drug_query.where(
+            DrugExposure.drug_concept_id == concept_id
+        )
+        self.measurement_query = self.measurement_query.where(
+            Measurement.measurement_concept_id == concept_id
+        )
+        self.observation_query = self.observation_query.where(
+            Observation.observation_concept_id == concept_id
+        )
+        return self
+
+    def add_age_constraint(
+        self,
+        left_value_time: str | None,
+        right_value_time: str | None
+    ) -> 'OMOPRuleQueryBuilder':
+        """
+        Apply age-at-event constraints to condition, drug, measurement, and observation queries.
+
+        Depending on which boundary is provided (left or right), this method applies a greater-than or less-than
+        comparator to filter records where the person's age at the event date satisfies the constraint.
+
+        Args:
+            left_value_time (str | None): Lower age bound as a string, or None if not specified.
+            right_value_time (str | None): Upper age bound as a string, or None if not specified.
+
+        Returns:
+            OMOPRuleQueryBuilder: The current instance with updated queries reflecting the age constraints.
+        """
+        if left_value_time is None or right_value_time is None:
+            return self
+        if left_value_time == "":
+            comparator = op.lt
+            age_value = int(right_value_time)
+        else:
+            comparator = op.gt
+            age_value = int(left_value_time)
+
+        self.condition_query = self._apply_age_constraint_to_table(
+            self.condition_query,
+            ConditionOccurrence.person_id,
+            ConditionOccurrence.condition_start_date,
+            comparator,
+            age_value,
+        )
+        self.drug_query = self._apply_age_constraint_to_table(
+            self.drug_query,
+            DrugExposure.person_id,
+            DrugExposure.drug_exposure_start_date,
+            comparator,
+            age_value,
+        )
+        self.measurement_query = self._apply_age_constraint_to_table(
+            self.measurement_query,
+            Measurement.person_id,
+            Measurement.measurement_date,
+            comparator,
+            age_value,
+        )
+        self.observation_query = self._apply_age_constraint_to_table(
+            self.observation_query,
+            Observation.person_id,
+            Observation.observation_date,
+            comparator,
+            age_value,
+        )
+        return self
+
+    def _apply_age_constraint_to_table(
+        self,
+        table_query: Select[Tuple[int]],
+        table_person_id: ClauseElement,
+        table_date_column: ClauseElement,
+        operator_func: Callable[[Any, Any], BinaryExpression[bool]],
+        age_value: int,
+    ) -> Select[Tuple[int]]:
+        """
+        Helper method to apply age constraints to a table query.
+
+        Args:
+            table_query: The table query to apply the age constraint to.
+            table_person_id: The person_id column in the table.
+            table_date_column: The date column in the table.
+            operator_func: The operator function to use in the constraint.
+            age_value: The age value to use in the constraint.
+
+        Returns:
+            The table query with the age constraint applied.
+        """
+        age_difference = SQLDialectHandler.get_year_difference(
+            self.db_manager.engine, table_date_column, Person.birth_datetime
+        )
+
+        constraint = operator_func(age_difference, age_value)
+
+        return table_query.where(
+            exists(
+                select(1).where(
+                    and_(
+                        Person.person_id == table_person_id,
+                        constraint,
+                    )
+                )
+            )
+        )
+
+    def add_temporal_constraint(
+        self,
+        left_value_time: str, 
+        right_value_time: str 
+    ) -> 'OMOPRuleQueryBuilder':
+        """
+        Adds a temporal constraint to OMOP queries relative to the current date,
+        using pre-parsed time values representing months.
+
+        Exactly one of `left_value_time` or `right_value_time` should be provided as
+        a numeric string (e.g., "6"), representing months. The other should be an
+        empty string.
+
+        The method filters events to either before or after the computed relative
+        date based on which time value is supplied:
+        - If `left_value_time` is given, events before (<=) that relative date are included.
+        - If `left_value_time` is empty, events after (>=) the `right_value_time` relative date are included.
+
+        Args:
+            left_value_time (str): Left-side time bound in months as a numeric string,
+                or empty string if unused.
+            right_value_time (str): Right-side time bound in months as a numeric string,
+                or empty string if unused.
+
+        Returns:
+            OMOPRuleQueryBuilder: The current instance with updated query filters.
+
+        Notes:
+        - This method assumes the input strings have already been parsed and
+          validated (e.g., "|6" converted to "6") before being passed in.
+        - The time values represent months relative to the current date.
+        """
+
+        if not left_value_time and not right_value_time:
+            raise ValueError(
+                "Temporal constraint requires exactly one time value. "
+                "Both left_value_time and right_value_time are empty."
+            )
+        
+        if left_value_time and right_value_time:
+            raise ValueError(
+                "Temporal constraint requires exactly one time value. "
+                f"Both values were provided: left='{left_value_time}', right='{right_value_time}'. "
+                "One must be an empty string."
+            )
+
+        if left_value_time == "":
+            time_value_supplied = right_value_time
+        else:
+            time_value_supplied = left_value_time
+
+        time_to_use = int(time_value_supplied)
+        time_to_use = time_to_use * -1
+
+        today_date = datetime.now()
+        relative_date = today_date + relativedelta(months=time_to_use)
+
+        if left_value_time == "":
+            self.measurement_query = self.measurement_query.where(
+                Measurement.measurement_date >= relative_date
+            )
+            self.observation_query = self.observation_query.where(
+                Observation.observation_date >= relative_date
+            )
+            self.condition_query = self.condition_query.where(
+                ConditionOccurrence.condition_start_date >= relative_date
+            )
+            self.drug_query = self.drug_query.where(
+                DrugExposure.drug_exposure_start_date >= relative_date
+            )
+        else:
+            self.measurement_query = self.measurement_query.where(
+                Measurement.measurement_date <= relative_date
+            )
+            self.observation_query = self.observation_query.where(
+                Observation.observation_date <= relative_date
+            )
+            self.condition_query = self.condition_query.where(
+                ConditionOccurrence.condition_start_date <= relative_date
+            )
+            self.drug_query = self.drug_query.where(
+                DrugExposure.drug_exposure_start_date <= relative_date
+            )
+        return self
+
+    def add_numeric_range(
+        self,
+        min_value: float = None,
+        max_value: float = None
+    ) -> 'OMOPRuleQueryBuilder':
+        """
+        Add numeric range constraints to measurement and observation queries.
+        
+        Args:
+            min_value: Minimum value (inclusive)
+            max_value: Maximum value (inclusive)
+            
+        Returns:
+            Self for method chaining
+            
+        Raises:
+            ValueError: If only one bound is provided or if min > max
+        """
+        if min_value is None and max_value is None:
+            return self
+        
+        if min_value is None or max_value is None:
+            raise ValueError(
+                "Both min_value and max_value must be provided for numeric range. "
+                f"Got min_value={min_value}, max_value={max_value}"
+            )
+        
+        min_val = float(min_value)
+        max_val = float(max_value)
+
+        if min_val > max_val:
+            raise ValueError(
+                f"min_value must be less than or equal to max_value. "
+                f"Got min_value={min_val}, max_value={max_val}"
+            )
+
+        self.measurement_query = self.measurement_query.where(
+        Measurement.value_as_number.between(min_val, max_val)
+        )
+        self.observation_query = self.observation_query.where(
+            Observation.value_as_number.between(min_val, max_val)
+        )
+
+        return self
+    
+    def add_secondary_modifiers(self, secondary_modifiers: list[int]) -> 'OMOPRuleQueryBuilder':
+        """
+        Filter the condition query by condition_type_concept_id values.
+
+        Adds an OR-combined filter to `condition_query` so that only condition
+        occurrences whose `condition_type_concept_id` matches one of the given
+        secondary modifier IDs are included. Has no effect on other table queries.
+
+        Args:
+            secondary_modifiers (list[int]): List of `condition_type_concept_id` values
+                to filter by. If empty or None, no filter is applied.
+
+        Returns:
+            OMOPRuleQueryBuilder: The current instance for method chaining.
+        """   
+        if not isinstance(secondary_modifiers, list):
+            raise TypeError(f"Expected list[int], got {type(secondary_modifiers).__name__}")
+
+        if any(not isinstance(mod, int) for mod in secondary_modifiers):
+            raise TypeError("All secondary modifier IDs must be integers")
+        
+        if not secondary_modifiers:
+            return self
+
+        modifier_constraints = [
+            ConditionOccurrence.condition_type_concept_id == modifier_id
+            for modifier_id in secondary_modifiers if modifier_id
+        ]
+
+        if modifier_constraints:
+            self.condition_query = self.condition_query.where(or_(*modifier_constraints))
+
+        return self
+
+    def build(self) -> CompoundSelect:
+        """
+        Build UNION query across all tables for this rule.
+
+        Returns:
+            CompoundSelect: UNION of all table queries
+        """
+        return union(
+            self.measurement_query,
+            self.observation_query,
+            self.condition_query,
+            self.drug_query
+        )
+
+
+class PersonConstraintBuilder:
+    """
+    Constructs SQLAlchemy filter constraints for querying the Person table
+    based on provided rules and concept mappings.
+    """
+
+    def __init__(self, db_manager: SyncDBManager):
+        self.db_manager = db_manager
+
+    def build_constraints(self, rule: Rule, concepts: dict[str, str]) -> list[ColumnElement[bool]]:
+        """
+        Generate SQL constraints for a given person-related rule.
+
+        Args:
+            rule (Rule): The rule defining the constraint parameters.
+            concepts (dict[str, str]): Mapping of concept IDs to their domains.
+
+        Returns:
+            List of SQLAlchemy boolean expressions representing the constraints.
+        """
+        if rule.varname == "AGE":
+            return self._build_age_constraints(rule)
+
+        concept_domain = concepts.get(rule.value)
+        if concept_domain == "Gender":
+            return self._build_gender_constraint(rule)
+        elif concept_domain == "Race":
+            return self._build_race_constraint(rule)
+        elif concept_domain == "Ethnicity":
+            return self._build_ethnicity_constraint(rule)
+
+        return []
+
+    def _build_age_constraints(self, rule: Rule) -> list[ColumnElement[bool]]:
+        """Build age range constraints."""
+        if rule.min_value is None or rule.max_value is None:
+            return []
+
+        age = SQLDialectHandler.get_year_difference(
+            self.db_manager.engine,
+            func.current_timestamp(),
+            Person.birth_datetime
+        )
+        return [
+            age >= rule.min_value,
+            age <= rule.max_value
+        ]
+
+    def _build_gender_constraint(self, rule: Rule) -> list[ColumnElement[bool]]:
+        """Build gender constraint."""
+        constraint = Person.gender_concept_id == int(rule.value)
+        return [constraint if rule.operator == "=" else ~constraint]
+
+    def _build_race_constraint(self, rule: Rule) -> list[ColumnElement[bool]]:
+        """Build race constraint."""
+        constraint = Person.race_concept_id == int(rule.value)
+        return [constraint if rule.operator == "=" else ~constraint]
+
+    def _build_ethnicity_constraint(self, rule: Rule) -> list[ColumnElement[bool]]:
+        """Build ethnicity constraint."""
+        constraint = Person.ethnicity_concept_id == int(rule.value)
+        return [constraint if rule.operator == "=" else ~constraint]
+
+
+class AvailabilitySolver():
 
     def __init__(self, db_manager: SyncDBManager, query: AvailabilityQuery) -> None:
         self.db_manager = db_manager
         self.query = query
+        self.person_constraint_builder = PersonConstraintBuilder(db_manager)
 
     @retry(
         stop=stop_after_attempt(3),
@@ -91,7 +462,31 @@ class AvailabilitySolver:
         (2) this function then iterates through the list of groups to resolve the logic (AND/OR) between groups
         """
         # resolve within the group
-        return self._solve_rules(results_modifier)
+        return self.solve_rules(results_modifier)
+
+    def solve_rules(self, results_modifiers: list[ResultModifier]) -> int:
+        """Main query resolution."""
+        concepts = self._find_concepts(self.query.cohort.groups)
+        low_number = self._extract_modifier(results_modifiers, "Low Number Suppression")
+        rounding = self._extract_modifier(results_modifiers, "Rounding")
+
+        with self.db_manager.engine.connect() as con:
+            group_queries = []
+
+            for group in self.query.cohort.groups:
+                group_query = self._build_group_query(group, concepts)
+                group_queries.append(group_query)
+
+            final_query = self._construct_final_query(
+                group_queries,
+                low_number,
+                rounding
+            )
+
+            output = con.execute(final_query).fetchone()
+            count = int(output[0]) if output is not None else 0
+
+        return apply_filters(count, results_modifiers)
 
     def _find_concepts(self, groups: list[Group]) -> dict[str, str]:
         """Function that takes all the concept IDs in the cohort definition, looks them up in the OMOP database
@@ -123,200 +518,155 @@ class AvailabilitySolver:
         }
         return concept_dict
 
-    def _solve_rules(self, results_modifier: list[ResultModifier]) -> int:
-        """Function for taking the JSON query from RQUEST and creating the required query to run against the OMOP database.
-
-        RQUEST API spec can have multiple groups in each query, and then a condition between the groups.
-
-        Each group can have conditional logic AND/OR within the group
-
-        Each concept can either be an inclusion or exclusion criteria.
-
-        Each concept can have an age set, so it is that this event with concept X occurred when
-        the person was between a certain age.
-
-        This builds an SQL query to run as one for the whole query (was previous multiple) and it
-        returns an int for the result. Therefore, all dataframes have been removed.
-
-        """
-        # get the list of concepts to build the query constraints
-        concepts: dict[str, str] = self._find_concepts(self.query.cohort.groups)
-
-        low_number: int = next(
+    def _extract_modifier(
+        self,
+        results_modifiers: list[ResultModifier],
+        result_id: str,
+        default_value: int = 10
+    ) -> int:
+        return next(
             (
                 item["threshold"] if item["threshold"] is not None else 10
-                for item in results_modifier
-                if item["id"] == "Low Number Suppression"
+                for item in results_modifiers
+                if item["id"] == result_id
             ),
-            10,
+            default_value,
         )
 
-        rounding: int = next(
-            (
-                item["nearest"] if item["nearest"] is not None else 10
-                for item in results_modifier
-                if item["id"] == "Rounding"
-            ),
-            10,
-        )
+    def _build_group_query(
+        self,
+        group: Group,
+        concepts: dict[str, str]
+    ) -> Union[Select[Tuple[int]], CompoundSelect]:
+        """Build query for a single group - a nested SQL expression."""
+        rule_table_queries = []
+        person_constraints = []
 
-        with self.db_manager.engine.connect() as con:
-            # this is used to store the query for each group, one entry per group
-            all_groups_queries: list[Union[Select[Tuple[int]], CompoundSelect]] = []
+        for rule in group.rules:
+            inclusion_criteria = rule.operator == "="
+            if rule.varcat == "Person":
+                constraints = self.person_constraint_builder.build_constraints(rule, concepts)
+                person_constraints.extend(constraints)
+            else:
+                rule_union = self._build_rule_query(rule)
+                rule_table_queries.append({
+                    'union_query': rule_union,
+                    'inclusion': inclusion_criteria
+                })
 
-            # iterate through all the groups specified in the query
-            for current_group in self.query.cohort.groups:
+        return self._construct_group_query(group, person_constraints, rule_table_queries)
 
-                # captures all the person constraints for the group
-                person_constraints_for_group: list[ColumnElement[bool]] = []
-                # captures all the rule table queries for the group
-                rule_table_queries: list[RuleTableQuery] = []
+    def _build_rule_query(self, rule: Rule) -> CompoundSelect:
+        """Build query for a single non-Person rule."""
+        builder = OMOPRuleQueryBuilder(self.db_manager)
 
-                # for each rule in a group
-                for current_rule in current_group.rules:
-                    # variables used to capture the relevant detail.
-                    # "time" : "|1:TIME:M" in the payload means that
-                    # if the | is on the left of the value it was less than 1 month
-                    # if it was "1|:TIME:M" it would mean greater than one month
-                    left_value_time: str | None = current_rule.left_value_time
-                    right_value_time: str | None = current_rule.right_value_time
-                    time_category: str | None = current_rule.time_category
+        if rule.value:
+            builder.add_concept_constraint(int(rule.value))
 
-                    # if the rule was not linked to a person variable
-                    if current_rule.varcat != "Person":
-                        # i.e. condition, observation, measurement or drug
-                        # NOTE: Although the table is specified in the query, to cover for changes in vocabulary
-                        # and for differences in RQuest OMOP and local OMOP, we now search all four main tables
-                        # for the presence of the concept. This is computationally more expensive, but is more r
-                        # reliable longer term.
-
-                        # initial setting for the four tables
-                        condition_query: Select[Tuple[int]] = select(
-                            ConditionOccurrence.person_id
-                        )
-                        drug_query: Select[Tuple[int]] = select(DrugExposure.person_id)
-                        measurement_query: Select[Tuple[int]] = select(
-                            Measurement.person_id
-                        )
-                        observation_query: Select[Tuple[int]] = select(
-                            Observation.person_id
-                        )
-
-                        """"
-                        RELATIVE AGE SEARCH
-                        """
-                        # if there is an "Age" query added, this will require a join to the person table, to compare
-                        # DOB with the data of event
-
-                        if (
-                            left_value_time is not None or right_value_time is not None
-                        ) and time_category == "AGE":
-                            condition_query, drug_query, measurement_query, observation_query = self._add_age_constraints(
-                                left_value_time, right_value_time, condition_query, drug_query, measurement_query, observation_query
-                            )
-
-                        """"
-                        STANDARD CONCEPT ID SEARCH
-                        """
-                        condition_query, drug_query, measurement_query, observation_query = self._add_standard_concept(
-                            current_rule, condition_query, drug_query, measurement_query, observation_query
-                        )
-
-                        """"
-                        SECONDARY MODIFIER
-                        """
-                        # secondary modifier hits another field and only on the condition_occurrence
-                        # on the RQuest GUI this is a list that can be created. Assuming this is also an
-                        # AND condition for at least one of the selected values to be present
-                        condition_query = self._add_secondary_modifiers(current_rule, condition_query)
-
-                        """"
-                        VALUES AS NUMBER
-                        """
-                        measurement_query, observation_query = self._add_range_as_number(
-                            current_rule, measurement_query, observation_query
-                        )
-
-                        """"
-                        RELATIVE TIME SEARCH SECTION
-                        """
-                        # this section deals with a relative time constraint, such as "time" : "|1:TIME:M"
-                        if (
-                            left_value_time is not None
-                            and right_value_time is not None
-                            and (left_value_time != "" or right_value_time != "")
-                            and time_category == "TIME"
-                        ):
-                            condition_query, drug_query, measurement_query, observation_query = self._add_relative_date(
-                                left_value_time, right_value_time, condition_query, drug_query, measurement_query, observation_query
-                            )
-
-                        """"
-                        PREPARING THE LISTS FOR LATER USE
-                        """
-                        # a switch between whether the criteria are inclusion or exclusion
-                        inclusion_criteria: bool = current_rule.operator == "="
-
-                        # Store the table queries for this rule to be used later in UNION
-                        # Union all table queries for this rule
-                        rule_union = union(measurement_query, observation_query, condition_query, drug_query)
-                        rule_table_queries.append({
-                            'union_query': rule_union,
-                            'inclusion': inclusion_criteria
-                        })
-
-                    else:
-                        """
-                        PERSON TABLE RELATED RULES
-                        """
-                        person_constraints_for_group = self._add_person_constraints(
-                            person_constraints_for_group, current_rule, concepts
-                        )
-
-                """
-                NOTE: all rules done for a single group. Now to apply logic between the rules
-                """
-                # Build the group query using UNION approach
-                group_query = self._construct_group_query(
-                    current_group,
-                    person_constraints_for_group,
-                    rule_table_queries
-                )
-
-                # Store the group query for later assembly
-                all_groups_queries.append(group_query)
-                logger.debug(f"Total groups stored: {len(all_groups_queries)}")
-
-            """
-            ALL GROUPS COMPLETED, NOW APPLY LOGIC BETWEEN GROUPS
-            """
-            # construct the query based on the OR/AND logic specified between groups using CTEs
-            final_query = self._construct_final_query(all_groups_queries, rounding)
-
-            if low_number > 0:
-                final_query = final_query.having(
-                    func.count() >= low_number
-                )
-
-            # here for debug, prints the SQL statement created
-            logger.debug(
-                str(
-                    final_query.compile(
-                        dialect=self.db_manager.engine.dialect,
-                        compile_kwargs={"literal_binds": True},
-                    )
-                )
+        if rule.left_value_time and rule.right_value_time and rule.time_category == "AGE":
+            builder.add_age_constraint(
+                left_value_time=rule.left_value_time,
+                right_value_time=rule.right_value_time
+            )
+        elif rule.left_value_time and rule.right_value_time and rule.time_category == "TIME":
+            builder.add_temporal_constraint(
+                left_value_time=rule.left_value_time,
+                right_value_time=rule.right_value_time
             )
 
-            output = con.execute(final_query).fetchone()
-            count = int(output[0]) if output is not None else 0
+        if rule.min_value is not None and rule.max_value is not None:
+            builder.add_numeric_range(rule.min_value, rule.max_value)
 
-        return apply_filters(count, results_modifier)
+        if rule.secondary_modifier:
+            builder.add_secondary_modifiers(rule.secondary_modifier)
+
+        return builder.build()
+
+    def _construct_group_query(
+        self,
+        current_group: Group,
+        person_constraints_for_group: list[ColumnElement[bool]],
+        rule_table_queries: list[RuleTableQuery]
+    ) -> Union[Select[Tuple[int]], CompoundSelect]:
+        """
+        Construct the query for a single group by processing inclusion/exclusion rules.
+
+        Args:
+            current_group: The group to construct a query for
+            person_constraints_for_group: Person-level constraints for this group
+            rule_table_queries: List of rule table queries for this group
+
+        Returns:
+            The constructed group query
+        """
+        # Build the group query using UNION approach
+        inclusion_queries: list[Union[Select[Tuple[int]], CompoundSelect]] = []
+        exclusion_queries: list[Union[Select[Tuple[int]], CompoundSelect]] = []
+
+        # Add person constraints as a separate query
+        if person_constraints_for_group:
+            person_query = select(Person.person_id).where(*person_constraints_for_group)
+            inclusion_queries.append(person_query)
+
+        # Add table queries for each rule
+        if rule_table_queries:
+            logger.debug(f"Processing {len(rule_table_queries)} rule table queries")
+            for i, rule_data in enumerate(rule_table_queries):
+                union_query = rule_data['union_query']
+                inclusion = rule_data['inclusion']
+                logger.debug(f"Rule {i}: inclusion={inclusion}")
+
+                if inclusion:
+                    # For inclusion: add the union directly
+                    inclusion_queries.append(union_query)
+                    logger.debug(f"Added inclusion query for rule {i}")
+                else:
+                    # For exclusion: store the union query to exclude people who match
+                    exclusion_queries.append(union_query)
+                    logger.debug(f"Added exclusion query for rule {i}")
+        else:
+            logger.debug("No rule table queries found")
+
+        # Create the final group query (without CTEs at this level)
+        if inclusion_queries:
+            if current_group.rules_operator == "AND":
+                # For AND logic, use INTERSECT which is more efficient than joins
+                group_query: Union[Select[Tuple[int]], CompoundSelect] = inclusion_queries[0]
+                for query in inclusion_queries[1:]:
+                    group_query = intersect(group_query, query)
+            else:
+                # For OR logic, use UNION
+                group_query = union(*inclusion_queries)
+        else:
+            # Start with all people if no inclusion queries
+            group_query = select(Person.person_id)
+
+        # Handle exclusion queries - remove people who match exclusion criteria
+        if exclusion_queries:
+            logger.debug(f"Processing {len(exclusion_queries)} exclusion queries")
+            try:
+                # Union all exclusion queries
+                exclusion_union = union(*exclusion_queries)
+                logger.debug("Exclusion union created successfully")
+
+                # Exclude people who match any exclusion criteria
+                exclusion_query = select(Person.person_id).where(
+                    ~Person.person_id.in_(select(exclusion_union.subquery()))
+                )
+                group_query = intersect(group_query, exclusion_query)
+
+                logger.debug("Exclusion queries processed successfully")
+            except Exception as e:
+                logger.error(f"Error processing exclusion queries: {e}")
+                raise
+
+        return group_query
 
     def _construct_final_query(
         self,
         all_groups_queries: list[Union[Select[Tuple[int]], CompoundSelect]],
-        rounding: int
+        rounding: int,
+        low_number: int
     ) -> Select[Tuple[int]]:
         """
         Construct the final query by applying OR/AND logic between groups using CTEs.
@@ -376,342 +726,18 @@ class AvailabilitySolver:
                 # Fallback to empty query
                 full_query_all_groups = select(func.count()).where(literal(False))
 
+        if low_number > 0:
+            full_query_all_groups = full_query_all_groups.having(
+                func.count() >= low_number
+            )
+
+        logger.debug(
+            str(
+                full_query_all_groups.compile(
+                    dialect=self.db_manager.engine.dialect,
+                    compile_kwargs={"literal_binds": True},
+                )
+            )
+        )
+
         return full_query_all_groups
-
-    def _construct_group_query(
-        self,
-        current_group: Group,
-        person_constraints_for_group: list[ColumnElement[bool]],
-        rule_table_queries: list[RuleTableQuery]
-    ) -> Union[Select[Tuple[int]], CompoundSelect]:
-        """
-        Construct the query for a single group by processing inclusion/exclusion rules.
-
-        Args:
-            current_group: The group to construct a query for
-            person_constraints_for_group: Person-level constraints for this group
-            rule_table_queries: List of rule table queries for this group
-
-        Returns:
-            The constructed group query
-        """
-        # Build the group query using UNION approach
-        group_queries: list[Union[Select[Tuple[int]], CompoundSelect]] = []
-        exclusion_queries: list[Union[Select[Tuple[int]], CompoundSelect]] = []
-
-        # Add person constraints as a separate query
-        if person_constraints_for_group:
-            person_query = select(Person.person_id).where(*person_constraints_for_group)
-            group_queries.append(person_query)
-
-        # Add table queries for each rule
-        if rule_table_queries:
-            logger.debug(f"Processing {len(rule_table_queries)} rule table queries")
-            for i, rule_data in enumerate(rule_table_queries):
-                union_query = rule_data['union_query']
-                inclusion = rule_data['inclusion']
-                logger.debug(f"Rule {i}: inclusion={inclusion}")
-
-                if inclusion:
-                    # For inclusion: add the union directly
-                    group_queries.append(union_query)
-                    logger.debug(f"Added inclusion query for rule {i}")
-                else:
-                    # For exclusion: store the union query to exclude people who match
-                    exclusion_queries.append(union_query)
-                    logger.debug(f"Added exclusion query for rule {i}")
-        else:
-            logger.debug("No rule table queries found")
-
-        # Create the final group query (without CTEs at this level)
-        if group_queries:
-            if current_group.rules_operator == "AND":
-                # For AND logic, use INTERSECT which is more efficient than joins
-                group_query: Union[Select[Tuple[int]], CompoundSelect] = group_queries[0]
-                for query in group_queries[1:]:
-                    group_query = intersect(group_query, query)
-            else:
-                # For OR logic, use UNION
-                group_query = union(*group_queries)
-        else:
-            # Start with all people if no inclusion queries
-            group_query = select(Person.person_id)
-
-        # Handle exclusion queries - remove people who match exclusion criteria
-        if exclusion_queries:
-            logger.debug(f"Processing {len(exclusion_queries)} exclusion queries")
-            try:
-                # Union all exclusion queries
-                exclusion_union = union(*exclusion_queries)
-                logger.debug("Exclusion union created successfully")
-
-                # Exclude people who match any exclusion criteria
-                group_query = select(Person.person_id).where(
-                    ~Person.person_id.in_(select(exclusion_union.subquery()))
-                )
-                logger.debug("Exclusion queries processed successfully")
-            except Exception as e:
-                logger.error(f"Error processing exclusion queries: {e}")
-                raise
-
-        return group_query
-
-    def _add_range_as_number(
-        self,
-        current_rule: Rule,
-        measurement_query: Select[Tuple[int]],
-        observation_query: Select[Tuple[int]]
-    ) -> tuple[Select[Tuple[int]], Select[Tuple[int]]]:
-        if current_rule.min_value is not None and current_rule.max_value is not None:
-            measurement_query = measurement_query.where(
-                Measurement.value_as_number.between(
-                    float(current_rule.min_value), float(current_rule.max_value)
-                )
-            )
-            observation_query = observation_query.where(
-                Observation.value_as_number.between(
-                    float(current_rule.min_value), float(current_rule.max_value)
-                )
-            )
-
-        return measurement_query, observation_query
-
-    def _add_age_constraints(
-        self, left_value_time: str | None, right_value_time: str | None,
-        condition_query: Select[Tuple[int]], drug_query: Select[Tuple[int]],
-        measurement_query: Select[Tuple[int]], observation_query: Select[Tuple[int]]
-    ) -> tuple[Select[Tuple[int]], Select[Tuple[int]], Select[Tuple[int]], Select[Tuple[int]]]:
-        """
-        This function adds age constraints to the query.
-        If the left value is empty it indicates a less than search.
-
-        Args:
-            left_value_time: The left value of the time constraint.
-            right_value_time: The right value of the time constraint.
-
-        Returns:
-            None
-        """
-        if left_value_time is None or right_value_time is None:
-            return condition_query, drug_query, measurement_query, observation_query
-
-        if left_value_time == "":
-            comparator = op.lt
-            age_value = int(right_value_time)
-        else:
-            comparator = op.gt
-            age_value = int(left_value_time)
-
-        condition_query = self._apply_age_constraint_to_table(
-            condition_query,
-            ConditionOccurrence.person_id,
-            ConditionOccurrence.condition_start_date,
-            comparator,
-            age_value,
-        )
-        drug_query = self._apply_age_constraint_to_table(
-            drug_query,
-            DrugExposure.person_id,
-            DrugExposure.drug_exposure_start_date,
-            comparator,
-            age_value,
-        )
-        measurement_query = self._apply_age_constraint_to_table(
-            measurement_query,
-            Measurement.person_id,
-            Measurement.measurement_date,
-            comparator,
-            age_value,
-        )
-        observation_query = self._apply_age_constraint_to_table(
-            observation_query,
-            Observation.person_id,
-            Observation.observation_date,
-            comparator,
-            age_value,
-        )
-
-        return condition_query, drug_query, measurement_query, observation_query
-
-    def _apply_age_constraint_to_table(
-        self,
-        table_query: Select[Tuple[int]],
-        table_person_id: ClauseElement,
-        table_date_column: ClauseElement,
-        operator_func: Callable[[Any, Any], BinaryExpression[bool]],
-        age_value: int,
-    ) -> Select[Tuple[int]]:
-        """
-        Helper method to apply age constraints to a table query.
-
-        Args:
-            table_query: The table query to apply the age constraint to.
-            table_person_id: The person_id column in the table.
-            table_date_column: The date column in the table.
-            operator_func: The operator function to use in the constraint.
-            age_value: The age value to use in the constraint.
-
-        Returns:
-            The table query with the age constraint applied.
-        """
-        age_difference = self._get_year_difference(
-            self.db_manager.engine, table_date_column, Person.year_of_birth
-        )
-
-        constraint = operator_func(age_difference, age_value)
-
-        # Use proper join instead of cross-join
-        return table_query.join(Person, Person.person_id == table_person_id).where(constraint)
-
-    def _get_year_difference(
-        self, engine: Engine, start_date: ClauseElement, year_of_birth: ClauseElement
-    ) -> ColumnElement[int]:
-        if engine.dialect.name in ("postgresql", "postgres"):
-            result = func.date_part("year", start_date) - year_of_birth
-            return result
-        elif engine.dialect.name == "mssql":
-            result = func.DATEPART(text("year"), start_date) - year_of_birth
-            return result
-        else:
-            logger.error(f"Unsupported database dialect: {engine.dialect.name}")
-            raise NotImplementedError(f"Unsupported database dialect: {engine.dialect.name}")
-
-    def _add_relative_date(self, left_value_time: str, right_value_time: str,
-        condition_query: Select[Tuple[int]], drug_query: Select[Tuple[int]],
-        measurement_query: Select[Tuple[int]], observation_query: Select[Tuple[int]]
-    ) -> tuple[Select[Tuple[int]], Select[Tuple[int]], Select[Tuple[int]], Select[Tuple[int]]]:
-        time_value_supplied: str
-
-        # have to toggle between left and right, given |1 means less than 1 and
-        # 1| means greater than 1
-        if left_value_time == "":
-            time_value_supplied = right_value_time
-        else:
-            time_value_supplied = left_value_time
-        # converting supplied time (in months) (stored as string) to int, and negating.
-        time_to_use: int = int(time_value_supplied)
-        time_to_use = time_to_use * -1
-
-        # the relative date to search on, is the current date minus
-        # the number of months supplied
-        today_date: datetime = datetime.now()
-        relative_date = today_date + relativedelta(months=time_to_use)
-
-        # if the left value is blank, it means the original was |1 meaning
-        # "I want to find this event that occurred less than a month ago"
-        # therefore the logic is to search for a date that is after the date
-        # that was a month ago.
-        if left_value_time == "":
-            measurement_query = measurement_query.where(
-                Measurement.measurement_date >= relative_date
-            )
-            observation_query = observation_query.where(
-                Observation.observation_date >= relative_date
-            )
-            condition_query = condition_query.where(
-                ConditionOccurrence.condition_start_date >= relative_date
-            )
-            drug_query = drug_query.where(
-                DrugExposure.drug_exposure_start_date >= relative_date
-            )
-        else:
-            measurement_query = measurement_query.where(
-                Measurement.measurement_date <= relative_date
-            )
-            observation_query = observation_query.where(
-                Observation.observation_date <= relative_date
-            )
-            condition_query = condition_query.where(
-                ConditionOccurrence.condition_start_date <= relative_date
-            )
-            drug_query = drug_query.where(
-                DrugExposure.drug_exposure_start_date <= relative_date
-            )
-
-        return condition_query, drug_query, measurement_query, observation_query
-
-    def _add_person_constraints(
-        self,
-        person_constraints_for_group: list[ColumnElement[bool]],
-        current_rule: Rule,
-        concepts: dict[str, str],
-    ) -> list[ColumnElement[bool]]:
-        concept_domain: str | None = concepts.get(current_rule.value)
-
-        if current_rule.varname == "AGE":
-            # AGE is a special case, as it is not a concept_id but a range.
-            min_value = current_rule.min_value
-            max_value = current_rule.max_value
-
-            if min_value is None or max_value is None:
-                return person_constraints_for_group
-
-            age = self._get_year_difference(
-                self.db_manager.engine, func.current_timestamp(), Person.year_of_birth
-            )
-            person_constraints_for_group.append(age >= min_value)
-            person_constraints_for_group.append(age <= max_value)
-
-        if concept_domain == "Gender":
-            if current_rule.operator == "=":
-                person_constraints_for_group.append(
-                    Person.gender_concept_id == int(current_rule.value)
-                )
-            else:
-                person_constraints_for_group.append(
-                    Person.gender_concept_id != int(current_rule.value)
-                )
-
-        elif concept_domain == "Race":
-            if current_rule.operator == "=":
-                person_constraints_for_group.append(
-                    Person.race_concept_id == int(current_rule.value)
-                )
-            else:
-                person_constraints_for_group.append(
-                    Person.race_concept_id != int(current_rule.value)
-                )
-
-        elif concept_domain == "Ethnicity":
-            if current_rule.operator == "=":
-                person_constraints_for_group.append(
-                    Person.ethnicity_concept_id == int(current_rule.value)
-                )
-            else:
-                person_constraints_for_group.append(
-                    Person.ethnicity_concept_id != int(current_rule.value)
-                )
-
-        return person_constraints_for_group
-
-    def _add_secondary_modifiers(self, current_rule: Rule, condition_query: Select[Tuple[int]]) -> Select[Tuple[int]]:
-        # Not sure where, but even when a secondary modifier is not supplied, an array
-        # with a single entry is provided.
-        secondary_modifier_list = []
-
-        for typeAdd in current_rule.secondary_modifier or []:
-            if typeAdd:
-                secondary_modifier_list.append(
-                    ConditionOccurrence.condition_type_concept_id == int(typeAdd)
-                )
-
-        if len(secondary_modifier_list) > 0:
-            condition_query = condition_query.where(or_(*secondary_modifier_list))
-
-        return condition_query
-
-    def _add_standard_concept(self, current_rule: Rule, condition_query: Select[Tuple[int]], drug_query: Select[Tuple[int]], measurement_query: Select[Tuple[int]], observation_query: Select[Tuple[int]]) -> tuple[Select[Tuple[int]], Select[Tuple[int]], Select[Tuple[int]], Select[Tuple[int]]]:
-        condition_query = condition_query.where(
-            ConditionOccurrence.condition_concept_id == int(current_rule.value)
-        )
-        drug_query = drug_query.where(
-            DrugExposure.drug_concept_id == int(current_rule.value)
-        )
-        measurement_query = measurement_query.where(
-            Measurement.measurement_concept_id == int(current_rule.value)
-        )
-        observation_query = observation_query.where(
-            Observation.observation_concept_id == int(current_rule.value)
-        )
-
-        return condition_query, drug_query, measurement_query, observation_query
